requirements.txt
blinker==1.6.2
Cerberus==1.3.4
click==8.1.3
Flask==2.3.1
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
PyJWT==2.6.0
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.0
